pyproject.toml
[project]
name = "open-cognition"
version = "0.1.0"
description = "AI-powered learning system with spaced repetition"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "surreal-basics>=0.3.0",
    "pydantic",
    "python-dotenv",
    "loguru",
    "jinja2",
    "fastmcp",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/open_cognition"]

[tool.hatch.build.targets.wheel.force-include]
"src/open_cognition/frontend/templates" = "open_cognition/frontend/templates"
"src/open_cognition/migrations" = "open_cognition/migrations"

[project.scripts]
open-cognition = "open_cognition.cli:main"
